Remove datapoint_constants. Make DatapointsQuery private. Rename type aliases
